Move internal crypto stuff from the header into the source file
